Phase 7 Security Hardening: fix 5 vulnerabilities found in full code audit
CRITICAL: database.rs derive_key() used random salt on every open, making
DB unopenable on restart. Fixed with deterministic salt from passphrase.

CRITICAL: FlareApplication hardware-backed key fallback used guessable
Build.FINGERPRINT. Fixed with encrypted challenge stored once.

HIGH: Rendezvous reply encryption used raw ephemeral public key (broadcast
in cleartext) as HKDF input — eavesdroppers could decrypt. Fixed with
proper X25519 ECDH between responder and querier ephemeral keys.

HIGH: signable_bytes() cast payload.len() to u16, truncating >65535 bytes.
Fixed: cast to u64.

MEDIUM: route_guard compute_max_allowed_ttl() ignored max_ttl_extension_factor
config, always returning 7-day hard cap. Fixed: infers original TTL and
applies 3.5x factor.

193 tests passing (2 new regression tests for key derivation determinism).

+++ flare-core/src/protocol/message.rs
@@ -97,7 +97,7 @@ impl MeshMessage {
         // ttl_seconds intentionally excluded — mutable by adaptive TTL
         bytes.extend_from_slice(&self.created_at_ms.to_le_bytes());
         bytes.push(self.content_type as u8);
-        bytes.extend_from_slice(&(self.payload.len() as u16).to_le_bytes());
+        bytes.extend_from_slice(&(self.payload.len() as u64).to_le_bytes());
         bytes.extend_from_slice(&self.payload);
         bytes
     }
